fixed the add reviews issue

+++ backend/GameDB_backend/src/main/java/devin/GameDB_backend/ReviewRepository.java
@@ -5,5 +5,5 @@ import java.util.List;
 
 public interface ReviewRepository extends JpaRepository<Review, Long> {
     List<Review> findByUser(User user);
-    List<Review> findByGame_GameId(Long gameId); // Ensure this method exists
+    List<Review> findByGame_GameId(Integer gameId); // Changed from Long to Integer
 }
